Fix two problems of cleaning up hub memberships
1. membership to clean may be empty
Example:
https://prow-gob.gcpnode.com/view/gcs/syllogi-gob-prow/logs/asm-security-gke-on-prem-hub-mcmn_istio_periodic/1433535395047411717
2021/09/02 21:01:42 ⚙️ gcloud container hub memberships delete  --quiet --project=tairan-asm-multi-cloud-dev
ERROR: (gcloud.container.hub.memberships.delete) argument (MEMBERSHIP : --location=LOCATION): Must be specified.

2. error when deleting a membership may be expected and should not return
Example:
https://prow-gob.gcpnode.com/view/gcs/syllogi-gob-prow/logs/asm-networking-gke-on-prem-mcmn_istio_periodic/1433535395047411713
2021/09/02 21:01:44 ⚙️ gcloud container hub memberships delete ci-tc47e37b25c24e6d --quiet --project=tairan-asm-multi-cloud-dev
ERROR: (gcloud.container.hub.memberships.delete) NOT_FOUND: Resource 'projects/tairan-asm-multi-cloud-dev/locations/global/memberships/ci-tc47e37b25c24e6d' was not found

+++ prow/asm/infra/deployer/tailorbird/tailorbird.go
@@ -109,9 +109,15 @@ func (d *Instance) Run() error {
 		}
 
 		for _, hm := range strings.Split(strings.TrimSpace(string(hms)), "\n") {
+			if strings.TrimSpace(hm) == "" {
+				// hm may be empty
+				continue
+			}
 			if err := exec.Run(fmt.Sprintf("gcloud container hub memberships delete %s --quiet --project=%s",
 				hm, onPremHubDevProject)); err != nil {
-				return err
+				// Error may be expected and should not cause the program to return, e.g., other test instances
+				// may also be cleaning up, which causes an error when deleting a membership that has been deleted.
+				log.Printf("Cleaning up %s returns an err: %v", hm, err)
 			}
 		}
 	}
